Extract SearchBar component
